+++ src/lib/util/constants.js
@@ -21,6 +21,7 @@ exports.DEFAULTS = {
 			warn: true,
 			wtf: true
 		},
+		createPiecesFolders: true,
 		disabledCorePieces: [],
 		language: 'en-US',
 		noPrefixDM: false,
@@ -95,9 +96,15 @@ exports.DEFAULTS = {
 				ignoreBlacklistedGuilds: true
 			},
 			providers: { enabled: true },
+			serializers: {
+				enabled: true,
+				aliases: []
+			},
 			tasks: { enabled: true }
 		},
-		schedule: { interval: 60000 }
+		schedule: { interval: 60000 },
+		slowmode: 0,
+		slowmodeAggressive: false
 	},
 
 	CONSOLE: {